Make a static-page dir and move modules into it.
