+++ frontend/lib/static-page/letter-static-page.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { StaticPage } from "./static-page";
-import { LambdaResponseHttpHeaders } from "../lambda/lambda";
-import { Route } from "react-router-dom";
+import { LambdaResponseHttpHeaders } from "../../lambda/lambda";
 
 type LetterStylesCss = {
   inlinePdfCss: string;
@@ -44,32 +43,3 @@ export const LetterStaticPage: React.FC<
     </html>
   </StaticPage>
 );
-
-type LetterStaticPageRouteInfo = ReturnType<
-  typeof createLetterStaticPageRouteInfo
->;
-
-export const createLetterStaticPageRouteInfo = (prefix: string) => ({
-  html: `${prefix}.html`,
-  pdf: `${prefix}.pdf`,
-});
-
-export function createLetterStaticPageRoutes(
-  routeInfo: LetterStaticPageRouteInfo,
-  render: (isPdf: boolean) => JSX.Element
-) {
-  return [
-    <Route
-      key={routeInfo.html}
-      path={routeInfo.html}
-      exact
-      render={() => render(false)}
-    />,
-    <Route
-      key={routeInfo.pdf}
-      path={routeInfo.pdf}
-      exact
-      render={() => render(true)}
-    />,
-  ];
-}
